fix(thermostat): resolve CI failures in MSCH schedule support
- Cast to size_t before multiplying two uint8_t constants in
  TestThermostatDelegateSchedules.cpp to fix a clang-tidy
  bugprone-implicit-widening-of-multiplication-result error.
- Shrink FakeThermostatDelegate's kMaxSchedules from 4 to 2 in
  TestThermostatClusterSchedules.cpp; each ScheduleStructWithOwnedMembers
  is ~1KB, and stack-constructing the fake blew the embedded
  -Werror=stack-usage budget on ESP32/nRF.
- Zero-initialize gDelegateTable with {} instead of { nullptr } so it
  still compiles when kThermostatEndpointCount is 0 (as reported by the
  new unit test's mock zap config).
- Skip building the new thermostat-server-schedule-test-srcs/tests GN
  targets under chip_build_controller_dynamic_server (Android JNI,
  Darwin framework, Python controller): that architecture skips
  zap-generated/gen_config.h entirely, so MATTER_DM_THERMOSTAT_CLUSTER_SERVER_ENDPOINT_COUNT
  is never defined there, and ThermostatCluster.cpp is never part of
  those builds' real dependency graph anyway.

+++ examples/thermostat/thermostat-common/tests/TestThermostatDelegateSchedules.cpp
@@ -117,7 +117,7 @@ TEST_F(TestThermostatDelegateSchedules, AppendPreservesExplicitHandle)
 TEST_F(TestThermostatDelegateSchedules, AppendFailsWhenPendingListIsFull)
 {
     auto & delegate            = ThermostatDelegate::GetInstance();
-    constexpr size_t kCapacity = kMaxNumberOfScheduleTypes * kMaxNumberOfSchedulesOfEachType;
+    constexpr size_t kCapacity = static_cast<size_t>(kMaxNumberOfScheduleTypes) * kMaxNumberOfSchedulesOfEachType;
 
     for (size_t i = 0; i < kCapacity; i++)
     {
